fix auto login on dapp
