+++ app/components/WrapHistory.vue
@@ -391,20 +391,6 @@ export default {
             this.fromWrapSelected = this.toWrapSelected
             this.toWrapSelected = temp2
         },
-        truncate (fullStr, strLen) {
-            if (fullStr.length <= strLen) return fullStr
-
-            const separator = '...'
-
-            let sepLen = separator.length
-            let charsToShow = strLen - sepLen
-            let frontChars = Math.ceil(charsToShow / 2)
-            let backChars = Math.floor(charsToShow / 2)
-
-            return fullStr.substr(0, frontChars) +
-               separator +
-               fullStr.substr(fullStr.length - backChars)
-        },
         async getTxs (coin) {
             try {
                 if (!coin) {
